+++ spec/language_spec.rb
@@ -46,7 +46,7 @@ module AcceptHeaders
         subject.new('en', 'us', q: 'a')
       end.must_raise Language::InvalidQError
 
-      e.message.must_equal 'invalid value for Float(): "a"'
+      e.message.must_match INVALID_FLOAT_PATTERN
 
       subject.new('en', 'us', q: '1')
     end