+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## HEAD
 
-  * Only strip accept param keys, values can contains white space if quoted.
+  * Support rbx invalid `Float` exception message.
+  * Only strip accept param keys, values can contain white space if quoted.
 
 ## 0.0.6 / November 17, 2014